[Refactor] Entity 폴더링 재작업 및 작명변경
