+++ Amor/Domain/Entity/Chat/ChatRequest.swift
