+++ Amor/Domain/Entity/Common/Empty.swift
@@ -7,4 +7,4 @@
 
 import Foundation
 
-struct Empty: Decodable { }
+struct Empty { }
