+++ Amor/Domain/Entity/Home/HomeChannelListContent.swift
