+++ Amor/Domain/Entity/Home/HomeDMListContent.swift
