+++ Amor/Domain/Entity/Home/HomeSectionModel.swift
