+++ Amor/Domain/Entity/User/Login.swift
@@ -7,26 +7,6 @@
 
 import Foundation
 
-/* Domain: 로그인 요청 모델 
- * Request -> DTO
- */
-struct LoginRequestModel {
-    let email: String
-    let password: String
-}
-
-extension LoginRequestModel {
-    func toDTO() -> LoginRequestDTO {
-        return LoginRequestDTO(
-            email: self.email,
-            password: self.password
-        )
-    }
-}
-
-/* Domain: 로그인 응답 모델 
- * DTO -> Response
- */
 struct Login {
     let user_id: String
     let nickname: String
@@ -38,15 +18,5 @@ struct Login {
         self.nickname = dto.nickname
         self.profileImage = dto.profileImage
         self.token = dto.token.toDomain()
-    }
-}
-
-struct Token {
-    let accessToken: String
-    let refreshToken: String
-    
-    init(_ dto: TokenResponseDTO) {
-        self.accessToken = dto.accessToken
-        self.refreshToken = dto.refreshToken
     }
 }
