+++ Amor/Domain/Entity/User/Profile.swift
@@ -1,50 +1,26 @@
 //
-//  ProfileElementEnum.swift
+//  MyProfile.swift
 //  Amor
 //
-//  Created by 김상규 on 11/8/24.
+//  Created by 김상규 on 11/4/24.
 //
 
 import Foundation
 
-enum Profile: String, CaseIterable {
-    case profileImage
-    case sesacCoin
-    case nickname
-    case phone
-    case email
-    case provider
-    case logOut
+struct Profile {
+    let email: String
+    let nickname: String
+    let profileImage: String?
+    let phone: String?
+    let provider: String?
+    let sesacCoin: Int
     
-    var name: String {
-        switch self {
-        case .profileImage:
-            "프로필 이미지"
-        case .sesacCoin:
-            "새싹 코인"
-        case .nickname:
-            "닉네임"
-        case .phone:
-            "연락처"
-        case .email:
-            "이메일"
-        case .provider:
-            "연결된 계정"
-        case .logOut:
-            "로그아웃"
-        }
-    }
-    
-    var placeholder: String {
-        switch self {
-        case .sesacCoin:
-            return "코인샵"
-        case .nickname:
-            return "닉네임을 입력해주세요"
-        case .phone:
-            return "연락처를 입력해주세요"
-        default:
-            return ""
-        }
+    init(_ dto: ProfileResponseDTO) {
+        self.email = dto.email
+        self.nickname = dto.nickname
+        self.profileImage = dto.profileImage
+        self.phone = dto.phone
+        self.provider = dto.provider
+        self.sesacCoin = dto.sesacCoin
     }
 }
