--- Amor/Domain/Entity/User/ProfileItem.swift
@@ -1,13 +1,0 @@
-//
-//  ProfileItem.swift
-//  Amor
-//
-//  Created by 김상규 on 11/7/24.
-//
-
-import Foundation
-
-struct ProfileItem: Hashable {
-    var profile: Profile
-    var value: String?
-}
