+++ Amor/Domain/Search/SearchResult.swift
