+++ Amor/Domain/Search/SearchSectionModel.swift
